simplify tab loading and move load/unload functions into browsertab
